+++ test/content/_njk-custom/filters.js
@@ -1,0 +1,5 @@
+const filters = (njkEnv) => {
+  njkEnv.addFilter('filterWithJustOneOutput', () => 'This is from the filter with just 1 output');
+}
+
+export default filters;